notas_conceito para contexto de consulta
