+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioNotaTipoValor.cs
@@ -1,12 +1,5 @@
-﻿using SME.SGP.Dominio.Enumerados;
-using System;
-
-namespace SME.SGP.Dominio.Interfaces
+﻿namespace SME.SGP.Dominio.Interfaces
 {
     public interface IRepositorioNotaTipoValor : IRepositorioBase<NotaTipoValor>
-    {
-        NotaTipoValor ObterPorCicloIdDataAvalicacao(long cicloId, DateTime dataAvalicao);
-
-        NotaTipoValor ObterPorTurmaId(long turmaId, TipoTurma tipoTurma = TipoTurma.Regular);
-    }
+    { }
 }
