+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioNotaTipoValorConsulta.cs
@@ -1,0 +1,12 @@
+﻿using SME.SGP.Dominio.Enumerados;
+using System;
+
+namespace SME.SGP.Dominio.Interfaces
+{
+    public interface IRepositorioNotaTipoValorConsulta : IRepositorioBase<NotaTipoValor>
+    {
+        NotaTipoValor ObterPorCicloIdDataAvalicacao(long cicloId, DateTime dataAvalicao);
+
+        NotaTipoValor ObterPorTurmaId(long turmaId, TipoTurma tipoTurma = TipoTurma.Regular);
+    }
+}
